* added gitignore support for visual studio files * improved the data type usage to decrease program size a bit * changed external live-display using LEDs to reserved LEDs on the stripe (pink LEDs above goal)

+++ Boss.h
@@ -8,7 +8,7 @@ class Boss
     void Kill();
     bool Alive();
     int _pos;
-    int _lives;
+    byte _lives;
     int _ticks;
   private:
     bool _alive;
